@@ -2,7 +2,7 @@ package com.abobrinha.caixinha.sync;
 
 import android.app.IntentService;
 import android.content.Intent;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 
 public class HistorySyncIntentService extends IntentService {
