@@ -5,7 +5,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
 import android.net.Uri;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.abobrinha.caixinha.data.HistoryContract;
 import com.firebase.jobdispatcher.Constraint;
